Update project schema tests
