+++ tests/schemas/test_project_invalid_schema_cases.py
@@ -1,0 +1,7 @@
+from pytest_cases import parametrize
+
+
+class CaseInvalidAttr:
+    @parametrize(attr=["name", "uuid"])
+    def case_attr(self, attr: str) -> tuple[str, None]:
+        return attr, None
